requirement.txt
django==1.4.16
distribute==0.7.3
MySQL_python==1.2.4
django-mptt==0.5.5
pillow==2.1.0
django-tinymce==1.5.1
django-staticfiles==1.2.1
south==0.8.1
django-suit==0.2.3